关于故事分类

+++ yryj/src/com/yryj/dao/UserDao.java
@@ -5,11 +5,11 @@ import com.yryj.model.User;
 public interface UserDao {
 	public void save(User user);
 
-	public void delete(int id);
+	public void delete(long id);
 
 	public void update(User user);
 
-	public User find(int id);
+	public User find(long id);
 	
 	public User findByName(String name);
