+++ yryj/src/com/yryj/sercvice/UserManager.java
@@ -5,11 +5,11 @@ import com.yryj.model.User;
 public interface UserManager {
 	public void save(User user);
 
-	public void delete(int id);
+	public void delete(long id);
 
 	public void update(User user);
 
-	public User find(int id);
+	public User find(long id);
 	
 	public User checkLogin(User user);
